CORE-9238 Replace admin TemplateAttributeList with ExpansionPanels.
